install-requirements.txt
torch==1.11.0
pyarrow
h5py
shapely
geographiclib
geopandas
osmnx
